fix(update): workbench13 restructuring files
