+++ packages/disks/workbench13/cloud/components/Info.vue
@@ -1,56 +1,47 @@
 <template>
-  <div class="wb-disks-workbench13-document-editor-info">
+  <div class="wb-disks-workbench13-cloud-info">
     <atom-markdown :content="content" />
   </div>
 </template>
 
 <script>
 
-import { toRef } from 'vue';
 import AtomMarkdown from '@web-workbench/core/components/atoms/Markdown';
+
 import useWindow, { windowProps, windowEmits } from '@web-workbench/core/composables/useWindow';
-import contextMenu from '../../documentEditor/contextMenu';
-import { getDocumentModelValue } from '../../utils';
+import contextMenu from '../contextMenu';
 
 export default {
   components: {
     AtomMarkdown
   },
 
-  props: {
-    ...windowProps,
-    model: {
-      type: Object,
-      default () {
-        return { value: getDocumentModelValue() };
-      }
-    }
-  },
+  props: { ...windowProps },
   emits: [
     ...windowEmits
   ],
 
   setup (props, context) {
-    const model = toRef(props, 'model');
     const windowContext = useWindow(props, context);
-    windowContext.setContextMenu(contextMenu, { model: model.value });
+    windowContext.setContextMenu(contextMenu);
     return windowContext;
   },
 
   data () {
     return {
       content: [
-        '# Document Editor', 'Version: **1.0**  \nCreated by **Thorn-Welf Walli**'
+        '# Cloud', 'Connect to a Firebase storage.  \n', 'Version: **1.0**  \nCreated by **Thorn-Welf Walli**'
       ].join('\n')
     };
   }
+
 };
 
 </script>
 
 <style lang="postcss" scoped>
-.wb-disks-workbench13-document-editor-info {
-  min-width: 320px;
+.wb-disks-workbench13-cloud-info {
+  min-width: 240px;
   padding: calc(var(--default-element-margin) * 2);
 }
 </style>
