+++ packages/disks/workbench13/index.js
@@ -1,24 +1,20 @@
 import { filter } from 'rxjs';
 import { ipoint } from '@js-basics/vector';
-import { markRaw, reactive } from 'vue';
 
 import WbComponentsConsole from '@web-workbench/core/components/Console';
 import { ITEM_META } from '@web-workbench/core/classes/FileSystem/Item';
 import { SYMBOL } from '@web-workbench/core/utils/symbols';
-import themeWhiteContrast from '@web-workbench/core/themes/whiteContrast';
-import { WINDOW_POSITION } from '@web-workbench/core/classes/WindowWrapper';
 
-import {
-  CONFIG_NAMES, CONFIG_DEFAULTS, DEFAULT_FONT,
-  getDocumentModelValue,
-  DEFAULT_FONT_SIZE,
-  FONT_FAMILES
-} from './utils';
+import clockAction from './clock';
+import calculatorAction from './calculator';
+import cloudAction from './cloud';
+import documentEditorAction, { DEFAULT_FONT_SIZE, FONT_FAMILES, CONFIG_DEFAULTS as CONFIG_DEFAULTS_DOCUMENT_EDITOR } from './documentEditor';
+import documentReaderAction from './documentReader';
 
 import documentHelpContent from './document-help.md?raw';
 
 export default ({ core }) => {
-  core.config.setDefaults(CONFIG_DEFAULTS);
+  core.config.setDefaults(CONFIG_DEFAULTS_DOCUMENT_EDITOR);
 
   return {
     locked: true,
@@ -291,284 +287,3 @@ export default ({ core }) => {
     ]
   };
 };
-
-function clockAction (core) {
-  return async ({ modules }) => {
-    const executionResolve = core.addExecution();
-    const [
-      component
-    ] = await Promise.all([
-      import('./components/Clock').then(module => module.default)
-    ]);
-    modules.windows.addWindow({
-      title: 'Clock',
-      component,
-      componentData: {},
-      options: {
-        scale: false,
-        scrollX: false,
-        scrollY: false
-      }
-    }, {
-      group: 'workbench13Clock'
-    });
-    executionResolve();
-  };
-}
-
-function calculatorAction (core) {
-  return async ({ modules }) => {
-    const executionResolve = core.addExecution();
-    const [
-      component
-    ] = await Promise.all([
-      import('./components/Calculator').then(module => module.default)
-    ]);
-    modules.windows.addWindow({
-      title: 'Calculator',
-      component,
-      componentData: {},
-      options: {
-        scale: false,
-        scrollX: false,
-        scrollY: false
-      }
-    },
-    {
-      group: 'workbench13Calculator'
-    });
-    executionResolve();
-  };
-}
-
-function cloudAction (core) {
-  return async ({ modules }) => {
-    async function updateItems (model) {
-      model.items = (await core.executeCommand('cloudList -json')).map(markRaw);
-    }
-
-    function hasError (value) {
-      if (value instanceof Error) {
-        return core.executeCommand(`openDialog "${value.message}" --title="${value.code}"`);
-      }
-    }
-
-    const model = reactive({ actions: {}, id: null, items: [] });
-    model.actions.login = async (email, password, storageId) => {
-      await hasError(await core.executeCommand(`cloudAuth -login --email="${email}" --password="${password}" --storage="${storageId}"`));
-      return updateItems(model);
-    };
-    model.actions.logout = async (id) => {
-      await hasError(await core.executeCommand(`cloudAuth --storage="${id}" -logout `));
-      return updateItems(model);
-    };
-    model.actions.connect = async (id, apiKey, url) => {
-      await hasError(await core.executeCommand(`cloudMount "CD${id.replace(/^(CD|cd)/, '')}" --apiKey="${apiKey}" --url="${url}"`));
-      return updateItems(model);
-    };
-    model.actions.disconnect = async (id) => {
-      await hasError(await core.executeCommand(`cloudUnmount "${id}"`));
-      return updateItems(model);
-    };
-
-    const executionResolve = core.addExecution();
-
-    modules.files.fs.root.events.subscribe(({ name }) => {
-      switch (name) {
-        case 'addItem':
-        case 'removeItem':
-          return updateItems(model);
-      }
-    });
-    await updateItems(model);
-
-    const [
-      component
-    ] = await Promise.all([
-      import('./components/Cloud').then(module => module.default)
-    ]);
-
-    modules.windows.addWindow({
-      title: 'Cloud',
-      component,
-      componentData: { model },
-      options: {
-        scale: false,
-        scrollX: false,
-        scrollY: false
-      }
-    },
-    {
-      group: 'workbench13Cloud'
-    });
-    executionResolve();
-  };
-}
-
-// Document Editor & Reader
-
-function documentEditorAction (core) {
-  const windowsModule = core.modules.windows;
-  return async ({ modules }, path) => {
-    const executionResolve = core.addExecution();
-    const [
-      WbComponentsDocumentEditor,
-      WbComponentsDocumentEditorPreview
-    ] = await Promise.all([
-      import('./components/DocumentEditor').then(module => module.default),
-      import('./components/documentEditor/Preview').then(module => module.default)
-    ]);
-
-    let model = {
-      actions: {},
-      value: getDocumentModelValue(),
-      fsItem: null,
-      [CONFIG_NAMES.DOCUMENT_EDITOR_SHOW_PREVIEW]: core.config.get(CONFIG_NAMES.DOCUMENT_EDITOR_SHOW_PREVIEW)
-    };
-
-    if (path) {
-      const fsItem = markRaw(await modules.files.fs.get(path));
-      const value = Object.assign(model.value, getDocumentModelValue(), fsItem.data);
-      model = Object.assign(model, {
-        fsItem,
-        value
-      });
-    }
-
-    const editorWindow = modules.windows.addWindow({
-      title: 'Document Editor',
-      component: WbComponentsDocumentEditor,
-      componentData: { model },
-      options: {
-        scale: false,
-        scrollX: true,
-        scrollY: true,
-        center: false,
-        embed: true,
-        borderless: true
-      },
-      layout: {
-        size: ipoint(540, 360)
-      }
-    }, {
-      group: 'workbench13DocumentEditor',
-      full: true
-    });
-
-    Object.assign(model.actions, {
-      close: () => {
-        editorWindow.close();
-      },
-      focus: () => {
-        editorWindow.focus();
-      },
-      reset: () => {
-        model.value = getDocumentModelValue();
-        model.fsItem = null;
-      }
-    });
-
-    let previewWindow;
-    model.actions.togglePreview = (toggle = true) => {
-      if (toggle) {
-        previewWindow = modules.windows.addWindow({
-          title: 'Preview - Document Editor',
-          component: WbComponentsDocumentEditorPreview,
-          componentData: { model },
-          options: {
-            scale: false,
-            scrollX: true,
-            scrollY: true,
-            center: false,
-            close: false,
-            embed: true,
-            borderless: true
-          },
-          layout: {
-            size: ipoint(540, 360)
-          }
-        }, {
-          group: 'workbench13DocumentEditor',
-          active: false
-        });
-        window.requestAnimationFrame(() => {
-          windowsModule.contentWrapper.setWindowPositions(WINDOW_POSITION.SPLIT_HORIZONTAL, [
-            editorWindow, previewWindow
-          ]);
-        }, 0);
-      } else if (previewWindow) {
-        editorWindow.unfocus();
-        previewWindow.close();
-        window.requestAnimationFrame(() => {
-          windowsModule.contentWrapper.setWindowPositions(WINDOW_POSITION.SPLIT_HORIZONTAL, [
-            editorWindow
-          ]);
-          editorWindow.focus();
-        });
-      }
-    };
-
-    core.modules.screen.setTheme(themeWhiteContrast);
-
-    return new Promise((resolve) => {
-      executionResolve();
-      editorWindow.events.pipe(filter(({ name }) => name === 'close')).subscribe(() => {
-        if (previewWindow) {
-          previewWindow.close();
-        }
-        core.modules.screen.setTheme(null);
-        resolve();
-      });
-    });
-  };
-};
-
-function documentReaderAction (core) {
-  return async ({ modules }, path) => {
-    let fsItem; let model = reactive({
-      actions: {},
-      fsItem: null,
-      value: getDocumentModelValue(),
-      fontFamily: DEFAULT_FONT
-    });
-    if (path) {
-      fsItem = markRaw(await modules.files.fs.get(path));
-      const value = Object.assign(model.value, getDocumentModelValue(), fsItem.data);
-      model = {
-        ...model,
-        fsItem,
-        value
-      };
-    }
-    const executionResolve = core.addExecution();
-    const [
-      component
-    ] = await Promise.all([
-      import('./components/DocumentReader').then(module => module.default)
-    ]);
-
-    const window = modules.windows.addWindow({
-      title: 'Document Reader',
-      component,
-      componentData: { model },
-      options: {
-        scale: true,
-        scrollX: false,
-        scrollY: false
-      }
-    }, {
-      full: true,
-      group: 'workbench13DocumentReader'
-    });
-
-    Object.assign(model.actions, {
-      close: () => {
-        window.close();
-      },
-      focus: () => {
-        window.focus();
-      }
-    });
-    executionResolve();
-  };
-}
